+++ GMap.NET.Core/GMap.NET.Projections/MapyCZProjection.cs
@@ -44,19 +44,19 @@ namespace GMap.NET.Projections
 
       #region -- WGSToMapyCZ --
 
-      public int[] WGSToPP(double la, double lo)
+      public long[] WGSToPP(double la, double lo)
       {
          var utmEE = wgsToUTM(DegreesToRadians(la), DegreesToRadians(lo), 33);
          var pp = utmEEToPP(utmEE[0], utmEE[1]);
          return pp;
       }
 
-      static int[] utmEEToPP(double east, double north)
+      static long[] utmEEToPP(double east, double north)
       {
          var x = (Math.Round(east) - (-3700000.0)) * Math.Pow(2, 5);
          var y = (Math.Round(north) - (1300000.0)) * Math.Pow(2, 5);
 
-         return new int[] { (int) x, (int) y };
+         return new long[] { (long)x, (long)y };
       }
 
       double[] wgsToUTM(double la, double lo, int zone)
@@ -203,13 +203,13 @@ namespace GMap.NET.Projections
          var size = GetTileMatrixSizePixel(zoom);
          {
             var l = WGSToPP(lat, lng);
-            ret.X = (int) l[0] >> (20 - zoom);
-            ret.Y = size.Height - ((int) l[1] >> (20 - zoom));
+            ret.X = (long)l[0] >> (20 - zoom);
+            ret.Y = size.Height - ((long)l[1] >> (20 - zoom));
          }
          return ret;
       }
 
-      public override PointLatLng FromPixelToLatLng(int x, int y, int zoom)
+      public override PointLatLng FromPixelToLatLng(long x, long y, int zoom)
       {
          PointLatLng ret = PointLatLng.Zero;
 
@@ -227,7 +227,7 @@ namespace GMap.NET.Projections
 
       public override GSize GetTileMatrixSizeXY(int zoom)
       {
-         return new GSize((int) Math.Pow(2, zoom), (int) Math.Pow(2, zoom));
+         return new GSize((long)Math.Pow(2, zoom), (long)Math.Pow(2, zoom));
       }
 
       public override GSize GetTileMatrixSizePixel(int zoom)
@@ -238,13 +238,13 @@ namespace GMap.NET.Projections
 
       public override GSize GetTileMatrixMinXY(int zoom)
       {
-         int wh = zoom > 3 ? (3 * (int) Math.Pow(2, zoom - 4)) : 1;
+         long wh = zoom > 3 ? (3 * (long)Math.Pow(2, zoom - 4)) : 1;
          return new GSize(wh, wh);
       }
 
       public override GSize GetTileMatrixMaxXY(int zoom)
       {
-         int wh = (int) Math.Pow(2, zoom) - (int) Math.Pow(2, zoom - 2);
+         long wh = (long)Math.Pow(2, zoom) - (long)Math.Pow(2, zoom - 2);
          return new GSize(wh, wh);
       }
    }
