+++ GMap.NET.WindowsPresentation/GMap.NET.WindowsPresentation/GMapMarker.cs
@@ -286,10 +286,10 @@ namespace GMap.NET.WindowsPresentation
          if(Map != null)
          {
             GPoint p = Map.FromLatLngToLocal(Position);
-            p.Offset(-(int)Map.MapTranslateTransform.X, -(int)Map.MapTranslateTransform.Y);
+            p.Offset(-(long)Map.MapTranslateTransform.X, -(long)Map.MapTranslateTransform.Y);
 
-            LocalPositionX = p.X + (int)(Offset.X);
-            LocalPositionY = p.Y + (int)(Offset.Y);
+            LocalPositionX = (int)(p.X + (long)(Offset.X));
+            LocalPositionY = (int)(p.Y + (long)(Offset.Y));
          }
       }
